% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs0.R
\name{adapt_seq}
\alias{adapt_seq}
\title{Adaptive tempering helper function}
\usage{
adapt_seq(temp, temp0, N, mt, mut, lambdat, G, cnew, snew, y)
}
\arguments{
\item{temp}{current temperature}

\item{temp0}{previous temperature}

\item{N}{Monte Carlo sample size}

\item{mt}{current sample of mass, vector of size \code{N}}

\item{mut}{current sample of signal strength parameter, vector of size \code{N}}

\item{lambdat}{current sample of baclground paths, array of dimensions \code{N*length(G)}}

\item{G}{grid over mass range}

\item{cnew}{matrix of signal normalizing constants - dimensions are (size of grid \code{G})by(number of categories/channels,41)}

\item{snew}{matrix of signal widths, same dimensions as \code{cnew}}

\item{y}{vector of observations}
}
\value{
current temperature to attain an effective sample size of \code{N/2}
}
\description{
Adaptive tempering helper function
}
