% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calib_funs.R
\name{exclusion_thresh}
\alias{exclusion_thresh}
\title{Exclusion threshold function}
\usage{
exclusion_thresh(m, G, snew, cnew, covmat, lambda_mean, alpha2, M)
}
\arguments{
\item{m}{the mass of Higgs}

\item{G}{grid over which the signal is evaluated}

\item{snew}{matrix of signal widths, same dimensions as \code{cnew}}

\item{cnew}{matrix of signal normalizing constants - dimensions are (size of grid \code{G})by(number of categories/channels,41)}

\item{covmat}{covariance matrix for teh background prior}

\item{lambda_mean}{mean vector for the background prior}

\item{M}{number of iterations used to estimate the lower 5 percent quantile of the mass marginal posterior}

\item{lambda0}{initial backgorund value}

\item{y}{observations}
}
\value{
exclusion threshold for a given mass
}
\description{
Exclusion threshold function
}
\examples{
load('data/SMC_full_poly')
L = dim(post2$thetat)[2]
dd_theta = density(post2$thetat[,L])
theta = dd_theta$x[which.max(dd_theta$y)]
dd_tau = density(post2$taut[,L])
tau = dd_tau$x[which.max(dd_tau$y)]
beta = apply(post2$beta[,L,], 2, mean)
xG = (G  - min(G))/(max(G) - min(G))
covmat = Gausscor(xG, theta, tau)
lambda_mean = log(poly(beta, xG)) - tau/2
cg = floor(seq(1, length(G), length = 100))
Qex = sapply(seq(100,180,1), exclusion_thresh,  G[cg], snew, cnew, covmat[cg,cg], lambda_mean[cg], M = 1000)
}
