% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs0.R
\name{Gausscor}
\alias{Gausscor}
\title{Gaussian covariance function}
\usage{
Gausscor(x, theta, tau)
}
\arguments{
\item{x}{vector of points for which the covariance matrix is generated}

\item{theta}{correlation parameter}

\item{tau}{variance parameter}
}
\value{
A covariance matrix of dimension ,equal to the length of \code{x}
}
\description{
Gaussian covariance function
}
