% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs0.R
\name{logpost}
\alias{logpost}
\title{Log posterior density function}
\usage{
logpost(lambda, mass, mu_s, theta, tau, beta, y = y, G = G, temp, cnew,
  snew, p0 = 0.5, mu_var, back_mean = "B polynomial")
}
\arguments{
\item{lambda}{vector of background values, the same size of grid \code{G}}

\item{mass}{mass value}

\item{mu_s}{signal streangth parameter}

\item{theta}{correlation parameter of the background logGP prior}

\item{tau}{variance parameter of the background logGP prior}

\item{beta}{polynomial coefficients of the background mean of the background logGP prior}

\item{y}{vector of observations}

\item{G}{grid over mass range}

\item{temp}{temperature parameter}

\item{cnew}{matrix of signal normalizing constants - dimensions are (size of grid \code{G})by(number of categories/channels,41)}

\item{snew}{matrix of signal widths, same dimensions as \code{cnew}}

\item{p0}{prior probability of $mass(Higgs)=0$}

\item{mu_var}{prior variance of signal strength parameter}

\item{back_mean}{parametric form used to specify the mean of background logGP prior}
}
\value{
the value of log posterior density for a set of parameters
}
\description{
Log posterior density function
}
