% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs0.R
\name{poly}
\alias{poly}
\title{Bernstein Polynomial}
\usage{
poly(beta, x, n = 4)
}
\arguments{
\item{beta}{polynomila coefficients}

\item{x}{the point at which the polynomial is evaluated}

\item{n}{order of the polynomial}
}
\description{
Bernstein Polynomial
}
