% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calib_funs.R
\name{sigfun}
\alias{sigfun}
\title{Signal function}
\usage{
sigfun(m, G, snew, cnew)
}
\arguments{
\item{m}{the mass of Higgs}

\item{G}{grid over which the signal is evaluated}

\item{snew}{matrix of signal widths, same dimensions as \code{cnew}}

\item{cnew}{matrix of signal normalizing constants - dimensions are (size of grid \code{G})by(number of categories/channels,41)}
}
\description{
Signal function
}
