% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calib_funs.R
\name{simpost}
\alias{simpost}
\title{Mass marginal posterior density with the Laplace approximation}
\usage{
simpost(j, m, G, snew, cnew, covmat, lambda_mean)
}
\arguments{
\item{j}{index}

\item{m}{the mass of Higgs}

\item{G}{grid over which the signal is evaluated}

\item{snew}{matrix of signal widths, same dimensions as \code{cnew}}

\item{cnew}{matrix of signal normalizing constants - dimensions are (size of grid \code{G})by(number of categories/channels,41)}

\item{covmat}{covariance matrix for teh background prior}

\item{lambda_mean}{mean vector for the background prior}

\item{lambda0}{initial backgorund value}

\item{y}{observations}
}
\description{
Mass marginal posterior density with the Laplace approximation
}
