% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calib_funs.R
\name{simpost_IS}
\alias{simpost_IS}
\title{Mass marginal posterior density with the Laplace approximation with importance weights under "no Higgs" assumption}
\usage{
simpost_IS(j, p0, G, snew, cnew, covmat, lambda_mean)
}
\arguments{
\item{j}{index}

\item{p0}{prior mass on \code{m = 0}}

\item{G}{grid over which the signal is evaluated}

\item{snew}{matrix of signal widths, same dimensions as \code{cnew}}

\item{cnew}{matrix of signal normalizing constants - dimensions are (size of grid \code{G})by(number of categories/channels,41)}

\item{covmat}{covariance matrix for teh background prior}

\item{lambda_mean}{mean vector for the background prior}

\item{lambda0}{initial backgorund value}

\item{y}{observations}
}
\description{
Mass marginal posterior density with the Laplace approximation with importance weights under "no Higgs" assumption
}
