% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs0.R
\name{SMC}
\alias{SMC}
\title{SMC sampler}
\usage{
SMC(N = 1000, G, y, cnew, snew, p0 = 0.5, mu_var = 0.5,
  back_mean = "B polynomial")
}
\arguments{
\item{N}{Monte Carlo sample size}

\item{G}{grid over mass range}

\item{y}{vector of observations}

\item{cnew}{matrix of signal normalizing constants - dimensions are (size of grid \code{G})by(number of categories/channels,41)}

\item{snew}{matrix of signal widths, same dimensions as \code{cnew}}

\item{p0}{prior probability of mass(Higgs)=0}

\item{mu_var}{prior variance of signal strength parameter}

\item{back_mean}{parametric form used to specify the mean of background logGP prior}
}
\value{
A list of sequence of posterior samples for model parameters,
sequence of acceptance rates for all parameters (for monitoring),
sequence of effective sample sizes (ESS) and sequence of temperatures
}
\description{
SMC sampler
}
\examples{
load('data/Higgs_data')
G = H_data$grid
y = H_data$y
cnew = H_data$cnew
snew = H_data$snew
post = SMC(N = 1e03, G = G, y = y, cnew = cnew, snew = snew, p0 = .5, mu_var = .5, back_mean = 'B polynomial')
save(post, file = 'data/SMC_full_poly')
}
