% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs0.R
\name{temploglik}
\alias{temploglik}
\title{Tempered log likelihood function}
\usage{
temploglik(par, y = y, G = G, temp, cnew, snew)
}
\arguments{
\item{par}{vector of parameters including background, mass and signal strength}

\item{y}{vector of observations}

\item{G}{grid over mass range}

\item{temp}{temperature parameter}

\item{cnew}{matrix of signal normalizing constants - dimensions are (size of grid \code{G})by(number of categories/channels,41)}

\item{snew}{matrix of signal widths, same dimensions as \code{cnew}}
}
\value{
the tempered log likelihood for a set of parameters
}
\description{
Tempered log likelihood function
}
